ci: run e2e tests under deno

+++ e2e/deno/cli.sh
@@ -12,20 +12,26 @@ case "$method" in
         ;;
     "update")
         # unpublish all packages
-        rm -rf .jsr-data/gcs/modules/@mtcute/*
-        docker compose exec jsr-db psql registry -U user -c "delete from publishing_tasks;"
-        docker compose exec jsr-db psql registry -U user -c "delete from package_files;"
-        docker compose exec jsr-db psql registry -U user -c "delete from npm_tarballs;"
-        docker compose exec jsr-db psql registry -U user -c "delete from package_version_dependencies;"
-        docker compose exec jsr-db psql registry -U user -c "delete from package_versions;"
-        docker compose exec jsr-db psql registry -U user -c "delete from packages;"
+        if [ -d .jsr-data/gcs/modules/@mtcute ]; then
+            rm -rf .jsr-data/gcs/modules/@mtcute
+            docker compose exec jsr-db psql registry -U user -c "delete from publishing_tasks;"
+            docker compose exec jsr-db psql registry -U user -c "delete from package_files;"
+            docker compose exec jsr-db psql registry -U user -c "delete from npm_tarballs;"
+            docker compose exec jsr-db psql registry -U user -c "delete from package_version_dependencies;"
+            docker compose exec jsr-db psql registry -U user -c "delete from package_versions;"
+            docker compose exec jsr-db psql registry -U user -c "delete from packages;"
+        fi
 
         # publish all packages
         docker compose run --rm --build build all
         
         # clear cache
-        rm -rf $(deno info --json | jq .denoDir -r)/deps
-        rm deno.lock
+        if command -v deno &> /dev/null; then
+            rm -rf $(deno info --json | jq .denoDir -r)/deps
+        fi
+        if [ -f deno.lock ]; then
+            rm deno.lock
+        fi
         ;;
     "clean")
         docker compose down
@@ -35,7 +41,9 @@ case "$method" in
         docker compose down
         ;;
     "run")
-        source .env
+        if [ -f .env ]; then
+            source .env
+        fi
 
         if [ -n "$DOCKER" ]; then
             # running behind a socat proxy seems to fix some of the docker networking issues (thx kamillaova)
@@ -55,6 +63,13 @@ case "$method" in
     "run-docker")
         source .env
         docker compose run --rm --build test $@
+        ;;
+    "ci")
+        set -eaux
+        mkdir .jsr-data
+        ./cli.sh start
+        ./cli.sh update
+        docker compose run --rm --build test
         ;;
     *)
         echo "Unknown command"
